feat: create long paragraph component for reward page
